requirements.txt
paho-mqtt==1.6.1
prometheus-client==0.15.0
watchgod==0.8.2
pyserial
